Got initial training run working

+++ src/UTTTGame.py
@@ -173,7 +173,7 @@ class UTTTGame():
         """
         mini_board_states_with_cond = (self.mini_board_states == condition).astype(np.float32)
         boards_matrix = mini_board_states_with_cond.reshape(3, 3) # 3x3
-        return np.kron(boards_matrix, np.ones((3, 3))) # 9x9
+        return np.kron(boards_matrix, np.ones((3, 3), dtype=np.float32)) # 9x9
 
     def _check_3x3_state(self, board_slice):
         """
